+++ src/open_samus_returns_rando/samus_returns_patcher.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from mercury_engine_data_structures.file_tree_editor import OutputFormat
 
 from open_samus_returns_rando.debug import debug_custom_pickups, debug_spawn_points
+from open_samus_returns_rando.exefs_patches import create_exefs_patches
 from open_samus_returns_rando.files import files_path
 from open_samus_returns_rando.logger import LOG
 from open_samus_returns_rando.lua_editor import LuaEditor
@@ -17,7 +18,6 @@ from open_samus_returns_rando.misc_patches.elevators import patch_elevators
 from open_samus_returns_rando.misc_patches.final_boss import patch_final_boss
 from open_samus_returns_rando.misc_patches.spawn_points import patch_custom_spawn_points
 from open_samus_returns_rando.misc_patches.text_patches import add_spiderboost_status, apply_text_patches
-from open_samus_returns_rando.multiworld_integration import create_exefs_patches
 from open_samus_returns_rando.patcher_editor import PatcherEditor
 from open_samus_returns_rando.pickups.custom_pickups import patch_custom_pickups
 from open_samus_returns_rando.pickups.pickup import patch_pickups
@@ -153,7 +153,7 @@ def patch_extracted(input_path: Path, output_path: Path, configuration: dict) ->
         out_code = output_path.joinpath("code.bin")
         out_exheader = output_path.joinpath("exheader.bin")
 
-        # Create Exefs patches for multiworld
+        # Create Exefs patches for multiworld and other binary modifications
         LOG.info("Creating exefs patches")
         create_exefs_patches(
             out_code, out_exheader,
